Rename to `permission_handler_html`
