+++ permission_handler_html/example/web/manifest.json
