+++ permission_handler_html/pubspec.yaml
@@ -1,7 +1,7 @@
-name: permission_handler_web
-description: A new Flutter package project.
-version: 0.0.1
-homepage:
+name: permission_handler_html
+description: Permission plugin for Flutter. This plugin provides the web API to request and check permissions.
+version: 0.1.0
+homepage: https://github.com/baseflow/flutter-permission-handler
 
 environment:
   sdk: ">=3.0.5 <4.0.0"
@@ -22,14 +22,10 @@ dev_dependencies:
   build_runner: ^2.1.2
   test: ^1.24.4
 
-# For information on the generic Dart part of this file, see the
-# following page: https://dart.dev/tools/pub/pubspec
-
-# The following section is specific to Flutter packages.
 flutter:
   plugin:
     implements: permission_handler
     platforms:
       web:
         pluginClass: WebPermissionHandler
-        fileName: permission_handler_web.dart
+        fileName: permission_handler_html.dart
